+++ .eslintrc.yml
@@ -1,7 +1,9 @@
 env:
   browser: true
   es6: true
+  jest: true
 extends: 'eslint:recommended'
+parser: babel-eslint
 parserOptions:
   ecmaFeatures:
     experimentalObjectRestSpread: true
@@ -9,10 +11,11 @@ parserOptions:
   sourceType: module
 plugins:
   - react
+  - react-native 
 rules:
   indent:
     - error
-    - tab
+    - 2
   linebreak-style:
     - error
     - unix
@@ -23,4 +26,6 @@ rules:
     - error
     - always
   no-console:
+    - warn
+  no-unused-vars:
     - warn
