+++ App.js
@@ -2,28 +2,28 @@ import React, { Component } from 'react';
 import { StyleSheet, Text, View } from 'react-native';
 import { createStackNavigator } from 'react-navigation';
 
-import Login from './Login';
-import MainShell from './MainShell';
+import Login from './src/Login';
+import MainShell from './src/MainShell';
 
 const RootStack = createStackNavigator(
-	{
-		Main: MainShell,
-		Login: Login,
-	},
-	{
-		initialRouteName: 'Login',
-	}
+  {
+    Main: MainShell,
+    Login: Login,
+  },
+  {
+    initialRouteName: 'Login',
+  }
 );
 
 export default class App extends Component {
-	render() {
-		return <RootStack />;
-	}
+  render() {
+    return <RootStack />;
+  }
 }
 
 const styles = StyleSheet.create({
-	container: {
-		flex: 1,
-		backgroundColor: '#fff'
-	}
+  container: {
+    flex: 1,
+    backgroundColor: '#fff'
+  }
 });
